+++ playground/vscode-setting/vite.config.ts
@@ -1,12 +1,12 @@
 import react from '@vitejs/plugin-react'
 import { defineConfig } from 'vite'
-import { i18nDetector } from 'vite-plugin-i18n-ally'
+import { i18n } from 'vite-plugin-i18n-ally'
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [
     react(),
-    i18nDetector({
+    i18n({
       root: __dirname,
     }),
   ],