+++ src/plugin/utils/init-options.ts
@@ -1,21 +1,22 @@
-import { type I18nDetectorOptions } from '..'
+import { type I18nAllyOptions } from '..'
 import { I18nAllyVscodeSetting } from './I18nAllyVscodeSetting'
 import { debug } from './debugger'
 
-const DEFAULT_OPTIONS: I18nDetectorOptions = {
+const DEFAULT_OPTIONS: I18nAllyOptions = {
   localesPaths: ['./src/locales', './locales'],
   root: process.cwd(),
   namespace: false,
-  autoDetectI18nConfig: true,
+  useVscodeI18nAllyConfig: true,
 }
 
-function getDefaultOptions(options?: I18nDetectorOptions): I18nDetectorOptions {
+function getDefaultOptions(options?: I18nAllyOptions): I18nAllyOptions {
   if (options?.dotVscodePath !== undefined) {
     console.warn(`dotVscodePath is deprecated, please use 'root' instead`)
   }
 
-  if (options?.autoDetectI18nConfig) {
-    const stopAt = typeof options.autoDetectI18nConfig === 'object' ? options.autoDetectI18nConfig.stopAt : undefined
+  if (options?.useVscodeI18nAllyConfig) {
+    const stopAt =
+      typeof options.useVscodeI18nAllyConfig === 'object' ? options.useVscodeI18nAllyConfig.stopAt : undefined
     const i18nAlly = new I18nAllyVscodeSetting(options?.root || (DEFAULT_OPTIONS.root as string), stopAt).init()
 
     debug('i18n-ally config:', i18nAlly)
@@ -31,12 +32,12 @@ function getDefaultOptions(options?: I18nDetectorOptions): I18nDetectorOptions {
   return DEFAULT_OPTIONS
 }
 
-export function initOptions(options?: I18nDetectorOptions) {
+export function initOptions(options?: I18nAllyOptions) {
   return {
     ...getDefaultOptions({
       ...DEFAULT_OPTIONS,
       ...options,
     }),
     ...options,
-  } as Required<I18nDetectorOptions>
+  } as Required<I18nAllyOptions>
 }